src/backend/requirements.txt
coreapi
coreschema
asgiref
attrs
cffi
cryptography
Django==5.0.2
django-cors-headers
django-filter
djangorestframework
djangorestframework-simplejwt
drf-spectacular
gunicorn
inflection
jsonschema
jsonschema-specifications
packaging
pillow
psycopg2-binary
pycparser
PyJWT
python-dotenv
PyYAML
referencing
rpds-py
ruff
setuptools
simplejwt
sqlparse
typing
typing_extensions
tzdata
uritemplate
whitenoise
python-dateutil